workbook_pimpl::find_style: fix warning

+++ src/xlcpp-pimpl.h
@@ -107,7 +107,7 @@ public:
         auto ret = styles.emplace(args...);
 
         if (ret.second)
-            ret.first->num = styles.size() - 1;
+            ret.first->num = (unsigned int)(styles.size() - 1);
 
         return &(*ret.first);
     }
